Version: 2003-09-23 93p-
Sourced from ast-open-history version 2003-09-23 and:
21c2f78e669926ab2a1e61690ccccbd3a27d6f5c4fbe3817ab992d9728f47b6d  INIT.2003-07-24.tgz

+++ src/cmd/3d/3d.h
@@ -500,7 +500,7 @@ typedef struct
 extern State_t		state;
 
 extern int		bprintf(char**, char*, const char*, ...);
-extern int		bvprintf(char**, char*, const char*, va_list);
+extern ssize_t		bvprintf(char**, char*, const char*, va_list);
 extern void		calldump(char**, char*);
 extern void		callinit(void);
 extern int		checklink(const char*, struct stat*, int);
